refactor: split executor based on module resolution
